server/requirements.txt
uvicorn 
gunicorn 
fastapi 
pymongo 
scikit-learn 
cryptography 
python-dotenv 
pydantic 
numpy 
torchvision 
panda  
